Share the order sub-resource controllers between admin and seller
The six admin/seller controller pairs for fulfillments, deliveries, labels,
returns, exchanges and claims each ran the same workflows with the same
arguments, written out twice. The workflow logic now lives in seven concerns
both branches include, and each controller keeps only what genuinely differs:
its serializer, its preloads, and the scope every id in a payload resolves
through.

The scope stays in the controllers deliberately. No concern fetches an order,
a variant or a stock location — those arrive already fetched and authorized,
and a lookup that must narrow to one seller's records is an abstract method
that raises rather than defaulting to something wider.

The seller branch also gains the orders base controller the admin branch has,
so the parent order is fetched once. That replaces four hand-written copies
whose authorization only ever asked for read access, even on a refund.

Two defects the sharing surfaced: an explicitly empty item list on receive
meant "receive everything" rather than "receive nothing", and a failed seller
fulfil returned a different error shape from every sibling endpoint.
