+++ spree/api/app/controllers/spree/api/v3/admin/orders/fulfillments_controller.rb
@@ -3,7 +3,15 @@ module Spree
     module V3
       module Admin
         module Orders
+          # The parcels owed on an order.
+          #
+          # Shipping, cancelling and splitting live in the shared concern;
+          # what is here is what makes them the operator's — the store-wide
+          # scope every id in a payload resolves through, and the two actions
+          # only an operator has.
           class FulfillmentsController < BaseController
+            include Spree::Api::V3::Orders::FulfillmentActions
+
             scoped_resource :fulfillments
 
             before_action :set_resource, only: [:show, :update, :fulfill, :mark_delivered, :cancel, :split]
@@ -39,51 +47,10 @@ module Spree
               end
             end
 
-            # PATCH /api/v3/admin/orders/:order_id/fulfillments/:id
-            def update
-              with_order_lock do
-                result = Spree.fulfillment_update_workflow.call(
-                  fulfillment: @resource,
-                  fulfillment_attributes: permitted_params.to_h
-                )
-
-                if result.success?
-                  render json: serialize_resource(@resource.reload)
-                else
-                  render_result_error(result)
-                end
-              end
-            end
-
-            # PATCH /api/v3/admin/orders/:order_id/fulfillments/:id/fulfill
-            #
-            # Marks the fulfillment as fulfilled. Passing `items` ships only
-            # those quantities: they are split into a new fulfillment which is
-            # the one that ships and is returned, leaving the remainder open.
-            def fulfill
-              with_order_lock do
-                result = Spree.fulfillment_fulfill_workflow.call(
-                  fulfillment: @resource,
-                  items: items_for_fulfill,
-                  tracking: fulfill_params[:tracking],
-                  tracking_carrier: fulfill_params[:tracking_carrier],
-                  notify_customer: notify_customer?(fulfill_params[:notify_customer]),
-                  force: fulfill_params[:force].to_b
-                )
-
-                if result.success?
-                  render json: serialize_resource(result.value)
-                else
-                  render_result_error(result)
-                end
-              end
-            end
-
             # PATCH /api/v3/admin/orders/:order_id/fulfillments/:id/mark_delivered
             #
-            # Confirms the customer received the goods. Staff can record this
-            # by hand — a merchant with no carrier integration still needs a
-            # delivered state.
+            # Staff confirming the parcel arrived. Deliberately the operator's
+            # alone: that goods arrived is the buyer's word, not the sender's.
             def mark_delivered
               with_order_lock do
                 result = Spree.fulfillment_mark_delivered_workflow.call(
@@ -100,73 +67,49 @@ module Spree
               end
             end
 
-            # PATCH /api/v3/admin/orders/:order_id/fulfillments/:id/cancel
-            def cancel
-              with_order_lock do
-                result = Spree.fulfillment_cancel_workflow.call(fulfillment: @resource)
-
-                if result.success?
-                  render json: serialize_resource(result.value)
-                else
-                  render_result_error(result)
-                end
-              end
-            end
-
-            # PATCH /api/v3/admin/orders/:order_id/fulfillments/:id/split
-            def split
-              with_order_lock do
-                variant = current_store.variants.find_by_prefix_id!(params[:variant_id])
-                quantity = params[:quantity].to_i
-
-                stock_location = if params[:stock_location_id].present?
-                                   find_stock_location!(params[:stock_location_id])
-                                 else
-                                   @resource.stock_location
-                                 end
-
-                fulfilment_changer = @resource.transfer_to_location(variant, quantity, stock_location)
-
-                if fulfilment_changer.run!
-                  fulfillments = @order.reload.fulfillments
-                  render json: {
-                    data: fulfillments.map { |s| serialize_resource(s) }
-                  }
-                else
-                  render_validation_error(fulfilment_changer.errors)
-                end
-              end
-            end
-
             protected
-
-            def model_class
-              Spree::Shipment
-            end
 
             def serializer_class
               Spree.api.admin_fulfillment_serializer
             end
 
-            def parent_association
-              :shipments
-            end
-
-            # The serializer embeds both, and the tracking summary reads the
-            # first consignment.
             def collection_includes
               [:deliveries, { shipping_labels: { file_attachment: :blob } }]
             end
 
-            # State changes go through the dedicated `fulfill`/`cancel` member
-            # actions, not mass assignment. `tracking` and `tracking_carrier`
-            # are the one-parcel shortcut onto the primary delivery.
             def permitted_params
               params.permit(:tracking, :tracking_carrier, :selected_delivery_rate_id, :stock_location_id)
             end
 
+            private
+
+            # Forcing past a stock or payment guard is the operator's call.
+            def fulfill_permitted_keys
+              super + [:force]
+            end
+
+            def fulfill_workflow_options
+              { force: fulfill_params[:force].to_b }
+            end
+
+            # The whole catalogue, and any warehouse this staff member may
+            # see: an operator ships anything the store sells.
+            def variant_for_split
+              current_store.variants.find_by_prefix_id!(params[:variant_id])
+            end
+
+            def stock_location_for_split
+              return @resource.stock_location if params[:stock_location_id].blank?
+
+              find_stock_location!(params[:stock_location_id])
+            end
+
             def create_params
               @create_params ||= params.permit(:stock_location_id, :tracking, :delivery_method_id, :cost, :status, metadata: {}, items: [:item_id, :quantity])
+            end
+
+            def mark_delivered_params
+              @mark_delivered_params ||= params.permit(:delivered_at, :notify_customer)
             end
 
             def find_stock_location!(id)
@@ -187,31 +130,6 @@ module Spree
               return if create_params[:items].nil?
 
               create_params[:items].map do |item|
-                {
-                  line_item: @order.line_items.find_by_prefix_id!(item[:item_id]),
-                  quantity: Integer(item[:quantity].to_s, exception: false)
-                }
-              end
-            end
-
-            def fulfill_params
-              @fulfill_params ||= params.permit(:tracking, :tracking_carrier, :notify_customer, :force, items: [:item_id, :quantity])
-            end
-
-            def mark_delivered_params
-              @mark_delivered_params ||= params.permit(:delivered_at, :notify_customer)
-            end
-
-            # Only an explicit false suppresses the email — an omitted flag, and
-            # anything unparseable, keeps the historic behavior of sending it.
-            def notify_customer?(value)
-              !ActiveModel::Type::Boolean.new.cast(value).equal?(false)
-            end
-
-            def items_for_fulfill
-              return if fulfill_params[:items].nil?
-
-              fulfill_params[:items].map do |item|
                 {
                   line_item: @order.line_items.find_by_prefix_id!(item[:item_id]),
                   quantity: Integer(item[:quantity].to_s, exception: false)
